Add files via upload
Providing Source files for version/rewrite 3. There is a new version on the way (rewrite 4) which will provide seemless map walking, animations and layered graphics, screen draw time reduction, and other elements. It should be the final aspect of the visual part of the engine. After that the sound/music part will be worked and then I'll return to the visual part to add some basic UI elements for the core gameplay part. At that point I will then start creating the game's content.